// data/github/current/ui/packages/pull-request-viewer/components/activity/timeline-items/__generated__/ReviewDismissedEventViaCommitStoryQuery.graphql.ts
export type ReviewDismissedEventViaCommitStoryQuery$variables = {
  id: string;
};
export type ReviewDismissedEventViaCommitStoryQuery$data = {
  readonly reviewDismissedEvent: {
    readonly " $fragmentSpreads": FragmentRefs<"ReviewDismissedEvent_reviewDismissedEvent">;
  } | null | undefined;
};
export type ReviewDismissedEventViaCommitStoryQuery = {
  response: ReviewDismissedEventViaCommitStoryQuery$data;
  variables: ReviewDismissedEventViaCommitStoryQuery$variables;
};